ME_QCALC_V2_MODULE_OOP

+++ src/main/java/com/crio/qcalc/StandardCalculator.java
@@ -8,7 +8,7 @@ public class StandardCalculator {
     }
 
     public void add(int num1, int num2){
-        add(num1,num2);
+        add((double)num1, (double)num2);
     }
     
     public void add(double num1, double num2){
@@ -16,7 +16,7 @@ public class StandardCalculator {
     }
 
     public void subtract(int num1, int num2){
-        subtract(num1,num2);
+        subtract((double)num1, (double)num2);
     }
     public void subtract(double num1, double num2){
         result = num1-num2;
@@ -24,7 +24,7 @@ public class StandardCalculator {
 
 
     public void multiply(int num1, int num2){
-        multiply(num1,num2);
+        multiply((double)num1, (double)num2);
     }
     public void multiply(double num1, double num2){
         result = num1*num2;
@@ -32,7 +32,7 @@ public class StandardCalculator {
 
 
     public void divide(int num1, int num2){
-       divide(num1,num2);
+       divide((double)num1, (double)num2);
     }
     public void divide(double num1, double num2){
         result = num1/num2;
